feat: add Home and MyLists page base
