--- src/frontend/src/pages/Groups.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-class Groups extends React.Component {
-    render () {
-        return (
-            <div>My Groups</div>
-        );
-    }
-};
-
-export default Groups;
